refactor(expr): use the shared regex engine; relocate it to applets::regex
- expr's `:`/`match` now delegate to the shared Pike VM instead of its own
  ~230-line anchored BRE matcher (BRE syntax, anchored at start, returns match
  length or the \(..\) capture). All four regex applets now share one engine.
- Move the shared engine from applets/text/regex.rs to applets/regex.rs so it
  can be used from both text/ (grep/sed/awk) and misc/ (expr) without forcing
  the whole `text` module into an expr-only build. Gated on
  any(grep, sed, awk, expr); each applet builds standalone again.

POSIX 671/0, lib 693/0. Verified full + expr/grep/sed/awk-only builds.
