+++ src/applets/regex.rs
